+++ terra/src/main/scala/com/dafttech/terra/engine/gui/elements/ElementSlotInventory.scala
@@ -17,7 +17,7 @@ class ElementSlotInventory(p: Vector2, val assignedInventory: Inventory) extends
   def useAssignedItem(causer: Player, pos: Vector2, leftClick: Boolean): Boolean = {
     if (assignedStack != null && assignedStack.amount > 0 && causer.worldObj.time > cooldownTime) {
       if ((!leftClick && assignedStack.use(causer, pos))) {
-        setCooldownTime(causer.worldObj, (assignedStack.`type`.toGameObject.asInstanceOf[Item]).getNextUseDelay(causer, pos, leftClick))
+        setCooldownTime(causer.worldObj, (assignedStack.`type`.toGameObject.asInstanceOf[Item]).getNextUseDelay(causer, pos, leftClick).toFloat)
         return true
       }
     }
